feat: Convert from 'd' to cameraRo setup
This relieves some of the processing from the GPU and puts it on the
cpu.

Always being tied to the origin is a bad idea in case we want more
sophisticated camera movements (hint hint). Specifying the ray origin
and the transform matrix (for camera rotation) is better than just
specifying the 'd' and calculating the matrix in the GPU.

Eventually it would be good to consolidate to a single matrix instead
of a vector and a matrix.