feat: Add IQ style composite cos's & add gradient module from three repo
ig post 2017-05-10

Decided not to add more layers of cosines because it was getting 'too'
detailed that way. Things were lost on small resolution renders and
wouldn't show in instagram.

Also ended up disabling the renderer to allow just a single still to
render and be saved out as my post. I did an animation of it on my PC
and it was so slow that it wasn't worth animating. That render took a
LONG time so I didn't want to tweak the variables to make it better.

Also the PC vs Mac rendering had pretty dramatically different render
colors so the PC render was also not worth it in that respect.

I added gamma like pow to the background colors to keep them from
blowing out w/ the bloom. This seems to be a real problem with my bloom
technique. It seems to catch too much as bloom-able. Maybe I just need
more subtle lightness values.