feat: Add polar parametric object & rotated map functions
The parametric object SDF is created by thinking of the radius and
angle coordinates as defining a series of points that are approximated
by spheres. This creates a smooth line with the thickness the same as
the diameter of the sphere.

The rotated map function was created to repeat a map multiple times by
rotating it a given number of times around a circle.

ig post 2017-05-12