feat: Add ripple deformations & banded timing intervals
The 'banded' timing is created by having a global mod'ed time that is
the entire loop but then smoothstepping a beginning and end to a slice
of that time for then moving distance fields or creating the ripple
effect.

The concept for this post was different primitive shapes submerging and
re-emerging as a different shape. The liquid nature of the floor they
were submerging into didn't materialize until after I already animated
them to go below the floor. It just felt as though it needed a ripple
effect to make it feel realistic. After that I knew I needed to add
wavelets and 'texture' to the floor to mimic the ever changing face of
an actual sea.

ig post 2017-05-16