feat: Distort sphere w/ a different scale factor
This started out as an attempt to optimize cos domain distortions using
their derivatives to scale the end result for the distance. Dividing
the end result by the running derivative was the most useful but was
hardly faster and in some cases enlarged the object. I even had once
that the object was everywhere because the derivative was so small.

I will have to revisit this when I have a better understanding of how
the Hubbard-Douady potential and it's derivative can be used to enhance
a mandelbrot set / other fractals.