refact: Translate the 'between' method for wavelength to hue
This should be equivalent to what was used prior but also supports
other hue values by interpolating between a set of fixed values. This
is also useful for even less sampling than is checked such as checking
only RGB. Should not give different results since it is queried with
exact hues used as reference.

Also moved it out into a module of it's own.