feat: Add normals based coloring on second refraction surface
The coloring is based on the second 'hit' surface on the distance
field. The physical analogy to this is that the object itself is clear
or glass but is wrapped in material such that you can only see the
materials color looking through the glass.

The concept behind the normals based coloring in this piece was to use
the different axes as the mixing term for three different colors.
Instead I used the dot product of the normal with the three basis
vectors as the mixing term. Since the mixing terms didn't necessitate
that the origin color prior to mixing was not a contributing factor, I
used a 'generic' color ( in this case white ) to blend into the others.
The result was coloring that reminded me of candy.

ig post 2017-05-21