feat: Add cell index to 2d & 3d voronoi
The 3d voronoi is unverified in functioning. I ended up using only the
2d so didn't get a chance to test it.

The purpose of this post was to recreate the ferrofluid look. Voronoi
was what inspired it as I saw the voronoi distance added to the surface
based on only the y & z coordinates and thought it resembled ferrofluid
magnetic field shape.

ig post 2017-05-27