feat: Use single scene lookup & and refraction for exit point
There is no point to use the IOR dispersion technique twice since the
first dispersion point represents a single color or wavelength. This
optimization saves on computation requirements and allows the RGBCMYK
version of dispersion to be used more as it is not exponential anymore
but instead linear (roughly).

ig post 2017-07-21