feat: Add work using trig to manipulate a plane
Had some great fun with iridescence & dispersion. They can really go
well together in the right proportions. Not sure if something is
iridescent that it must also have dispersion. I would think it's not
necessarily true since iridescence comes from the microfacets being
smaller than the wavelength of the light at certain angles. Dispersion
is a difference in index of refraction between two mediums and requires
a depth to one of the mediums to make a significant difference. Maybe
an iridescent and dispersive material would be a glass covered metallic
painted surface.

ig post 2017-08-18