feat: Add moody lighting experiment w/ directional based mix warp
I wanted to have an object popping in and out of shadows using intense
fog. I then thought about using a dimension as the input for a mix
between a regular version of the object and a cosine warped version.
Doing this on the vertical axis would create a potentially drippy
looking effect.

In the end it looked like gallium hence the piece's name is Gallium.
Didn't receive many likes though. Probably because it's too dark.

ig post 2017-09-02