refact: Move rendering external to App so renderer can be headless etc
App is not responsible for rendering isnt it is bootstraped by `index.js`
which either hooks up a capturer or rigs it for running alone.

Having it based in WebVR rendering makes it very hard to separate.
Normally an App would be given a webgl context and then would render
itself to it but when it's WebVR, it is more involved and, really, the
App shouldn't be specifically geared towards a VR setup though will
have to support it. It could be that it is further corrected in that
the fragment shader (or whatever) takes camera parameters and can be
run multiple times with the camera parameters changed between frames if
rendering separate eyes. This is what is happening currently but it
doesn't have a native support for rendering with a simple camera etc,
without using the WebVRManager.