feat: Add red neon piece w/ black seascape
I had to adjust the bloom / glow filter to be additive instead of using
the render buffer alpha for blending between the bloom + base buffer &
the background.

I also attempted multiple reflection passes with each reflection ray
being adjusted with noise.. It was mildly successfully but tended to
have a multiple exposure like appearance instead of a blur of the
reflection.

Moved the light for once. It immediately because more cinematic once I
did. I should watch out for lightings effect on the scene from now on.

ig post 2017-10-06