feat: Loop over arbitrary hues to create dispersion color
Turns out that my red only versions were due to my scale being off for
the hue. Once fixed, I was able to create a loop of however many hues I
wanted to render.

There is still some 'banding' of certain colors no matter the number of
iterations. I am not sure why but maybe they are not sampling related.
Maybe they are due to the returned indexes of refraction for those hues.