feat: Create a piece using two textures, 1 dispersion & 1 noise
The dispersion piece is ray marched and uses a high quantity for the
number of hues.

The purpose of this piece was to exercise my ability to create
procedural textures and create different but complementary textures.

I also used the classic cosine rainbow palette from IQ instead of the
hsv() function.

ig post 2017-11-05