feat: Create piece using mod as a 'random' source of rotation
mod() can be used to create random values by using prime numbers or
numbers with large gcd to limit the period or 'orderliness' of the
output values. I do this by multiplying the input by a primeish number
and then setting the modulo to be another primeish number. This is
likely related to other noise functions but is especially useful for
non-smooth noise functions as it is cheap.

ig post 2017-11-13