feat: Add grid of rotating plus signs driven by noise
I came up with another technique for looping noise. Instead of having
an interpolation between two noise values (Connor Bell's technique),
what if the input was offset just by a polar/spherical coordinate that
has a period. This will enable the detail of a function to be preserved
and never averaged but does potentially mean that the circular motion
will be noticed. In this example it's solved by the fact that the noise
values are displayed relatively far apart and at a slow enough pace
that the noise space translation isn't apparent.

ig post 2017-11-30