feat: Experiment with color layering techniques
I chose to distort the grid like result to add more interest but the
original intent was to create colored glass that had a kaleidoscopic
like feel to it in that it comprised of multiple layers of colored
glass that would tumble over each other.

I created 3 versions of the piece.

1. Is a distorted but asymmetric
form.
2. Is a symmetric form created by adding an `abs` before any warps are
applied.
3. Is a 3x slowed down version of #2. This was to allow viewers to see
more detail about the shapes as they evolve.