feat: Experiment w/ composite rotated vector coloring technique
I decided that a good way to allow different continuous values to
affect a relatively set brightness color would be applying a series of
rotation matrices on a vector with a set length.

These rotations mean that it is smooth and continuous. I thought I
would set the axis for each parameter as just on of the XYZ axes, but
before even trying that, I decided that normalized vectors with equal
values in XY, YZ & XZ would be more interesting and complex.

ig post 2017-12-12