feat: Correct UVs in background & create distorted grid
I started out trying to create a white version of the previous daily
which was focused on dispersion. It didn't really turn out. I felt that
it was either too subtle or that I had to go so dark that it ended up
looking effectively the same. So I switched to a previous piece I had
attempted, but this time I tried it as a 2D shader.

While working on this daily, I realized that the UVs sent to the
background while raymarching were not aligned with the UVs sent to the
background while being rendered in the final pass. The raymarching
background was using the FragCoord vec2 varying and the final pass was
just using its own UVs that it derived. I have changed it to be
consistent.

ig post 2017-12-16