feat: Create cellular tile plan of iridescent / thin film gems
I may have gained a sense of understanding from this experimentation.
A problem I frequently run into is the fact that the 'dark' regions are
actually a washed out gray. This is because these areas do contain bits
of color on the initial entry into the medium, but then die out
quickly.

I tried making the color a single HSB / rainbow lookup up when the hue
being an accumulation of the hues of colors that would have been added
together into a muddled gray, but the depths was still missing. It was
harder to get the sense of depth from this technique.

Future things to try:
- not using nsin() as a domain wrapper in the `lum` addition value
- Another technique for domain warping in the holographic loop besides
  a dot product.
- Experiment further with the lambertian dot product modifying the hue
  of a color.

ig post 2017-12-29