feat: Play w/ volumetric textures & vfbm warp noise
I am trying to create more depth in the volumetric textures. This
attempt was mildly success. I modified how the volume colors are
blended by using IQ's technique in hell (link below). It helped when
the texture was monochromatic but once I added color, it was a balance
between color range and noisiness. It was hard to differentiate colors
once there were more of them.

https://www.shadertoy.com/view/MdfGRX

ig post 2017-12-30