feat: Created a holographic material like the aurora borealis
I figured out something that helps to increase the depth while
maintaining clarity. Just cut out the 'low' end of the value range for
a given position. This means more solid dark areas that allow lower
brighter values to shine through!

I have a bit of banding in this material but I left it because of the
texture it creates.

ig post 2018-01-02