feat: Create a lemniscate shape piece
I was attempting to have the lemniscate version of a mobius strip but
because I had a mirror the left and right, the mobius strip twist
didn't work. It probably could be tricked with a little bit of clever
angle manipulation based on which side of the mirroring you are on.

ig post 2018-03-23