feat: Create a procedural ripping animation
I used noise to emulate various parts of the material as it rips. First
I used voronoi to create the 'strands' between the two sides. These
were only mildly successfully as they didn't stay as long as I was
hoping for. Second, I used noise to create flecks of material. Finally
I used long period noise to warp the edge of the two sides of the
material into an irregular pattern.

ig post 2018-05-22