feat: Create cutout w/ noise
Played with creating an undulating grid with colorful lights. That
didn't work out how I envisioned it. So I started experimenting and
found that I liked the look of a sphere cut out of a box. But I
couldn't make it interesting enough, so I was trying to spice it up by
added noise during a transition. Because I was playing with noise, I
saw what it looked like to cutout noise from the box and decided to go
with that with a red & cyan pair of lights.

ig post 2018-09-29