feat: Recreate 'Movement in Squares' by Bridget Riley
Not quite correct because her scale of the squares as they get squished
is not exponential. I also think there is an issue with how I scale the
x coordinate. It doesn't seem to slope evenly straight "up". It's as
though the top of the scale range dips for a bit then finally goes
down. This could be also caused by a sampling issue where each pixel is
just getting the same color because the domain is scaled so much.

ig post 2019-01-30