feat: Create a twitching twisted cube
I was experimenting with new ways to morph and distort the space. But
then was surprised with a sudden movement that felt like a twitch from
the object. Still not sure why the movement is so sudden. Maybe its
because I'm moving the axis so quickly but yet the values for x, y & z
are pretty close to each other (being only 2x radius apart) so when the
`cos` hits `0` for x, y is not far behind etc. That means that the axis
becomes very close to `vec3(0)` for a second and that any difference
between the different dimensions will be huge in the normalized axis
version of it.

ig post 2019-06-10