feat: Create a warm color iridescent / metallic material
I came into this piece wanting to create something bright and happy.

I was surprised to see that the orthographic camera projection changed
how the material presented. There was more nuance in the ortho version
which is likely due to little or no change in the camera ray direction.

ig post 2019-06-12