feat: Change the center luminance over time to force movement
I knew I wanted to have the highlights and shadows change more than
swapping places, so I added the ability to 'rotate' where they would be.

As part of this process, I decided to hook up a dat.GUI parameter to the
center color. I was surprised how much the visual looked liked it moved
when I slid the color back and forth.

Might make this a website that allows you to play with the Fraser-Wilcox
illusion.

ig post 2019-09-12