feat: Play more with the sliced up cropped layers
Tried to be more adventurous with my background color and cosine
palette. I don't like it as much as yesterday and was hoping that I'd
set a background color and just 'find' a good setting for the cosine
palette. Still, the cosine palette & dispersion is this lovely metallic
purple, orange and deep blue color.

ig post 2020-03-04