feat: Render scene in the past / alt universe based on material id
Setting up the renderer to support this would have been much easier if
the global state wasn't a mess, glslify would have allowed me to use the
global state in a required module (which I thought it did but then all
the variables were renamed etc), and finally if I could use recursion in
glsl. This may be possible in webgl 2.0.

Refactored time variables into an include as the final pass shader had
old values. I should refactor render time to be set in a single
location.

ig 2020-06-08