feat: Create accidental flower blob w/ peachy hues
I was trying to create liquid drops that swirled that quickly started
looking like a Yin & Yang. I wanted to move away from that so I tried
adding a sphere that the drops would cut out from, but forgot to reset
the domain to before the polar modulus and then created the blobby
flower.

ig 2020-07-09