feat: Quantize the hue of a dispersion simulation
The hue is quantized in hsv space and converted back to RGB using a
cosine palette. So this is not a 1-to-1 hue of the dispersion simulation
result, just an approximation. Especially since I scaled the hue value
too. Finally I mix in a bit of the value coordinate by mixing it towards
white and multiplying the quantized hue color.

ig 2020-07-19