feat: Warp grid centers only
I finally figured out a way to warp the centers of the grid without
warping the shapes in that cell. I just needed to not use the position
in the cell to offset the center. AKA if you are doing a layered cosine
warp, don't use the position in the `cos()` function. Instead use the
cell index which changes per cell so will be uniform across the cell.

ig 2020-07-22