feat: Play w/ twisting 4D spheres
The surface texture from applying the cosine warp and two 3D twists on
different sets of coordinates of the 4D space created a more lovely and
organic feel. This is probably because of the 'hidden' nature of that
extra dimension's role in the warping of the shape. Maybe the brains
ability to pattern match is dulled when it reaches out of our realm of
existence, so instead it sees pseudo-patterns that have rules but they
are hidden or too complex just like nature as a whole.

ig 2020-08-18