feat: Play w/ shifting cells around
I wonder how a tile travels in this version. There are tiles that just
move in a square as the end up always being included in the movement.
But what about the other tiles. The top left of a quad seems to be
constantly being pushed to the left and right.

ig 2020-08-27