feat: Stumble upon a weird material that 'shadow reflects'
I know 'shadow reflects' doesn't make sense and is illogical. But the
way that the darkness in the material moves makes it feel like its
reflecting a shadow. Maybe from the room or inside it. Really there is
no real world analogy because it uses the orbit trap to otherwise just
nudge a pretty iridescent reflective surface.

ig 2020-11-14