fix: Render w/o a delay
Originally I thought that the delay was necessary because there seemed
to be a race condition in the CCapture.js library for the filename of
the frame's file that was added to the tar being generated. Turns out
that it was caused by me replacing the timing functions (setTimeout,
requestAnimationFrame, etc) to the original functions.

I am not sure why I originally restored these functions though I know it
had to do with the webVR implementation I originally used. Currently I
do not support a webVR renderer, but if I do in the future, I'll have to
revisit this fix in the future.