feat: Create below 0 black base color w/ dispersion highlights
I found it weird I couldn't make it as black as I wanted. I've always
struggled with that given black objects still have Fresnel and specular
lighting which then adds to the black to make a grey. Setting the
'black' base to a negative number (and removing the guards against it)
allowed for more range... or so I like to think of it.

ig 2020-12-23