feat: Create an interesting dispersion material
Really just tweaking some settings and thinking about how to improve the
dispersion simulation. One major item is the fact that it doesn't really
look at the scene when it 'exits' the object for the two step version.
This means when the majority of the object could be dark because of the
dispersion simulation when it should generally be a bright object with a
bright background etc. Something to think about and improve.

ig 2021-03-17