feat: Obliterate a circle into small pieces
Combined the mask and space modifiers for the split operator. To keep
the interface cleaner I used an `inout` param for the mask. This way
someone could use the function without needing to separate out the
response into the space and then the mask etc.

I realized that the `start` split parameter is not the coordinate in the
absolute space but in the modified space. This has two issues:
1. the parameter needs to account for previous modifications to the
   space which can get really complex. An example of a hard set of
   params would be splitting the object into a grid. This isn't the
   naive "set start at regular intervals" solution.
2. the second is that when the animation of the split overlaps with
   another, it is adjusting each other's start position causing the
   split location to 'slip'. If the start position was based on absolute
   space, it would not cause on split to affect the position of another
   split.

Not sure how I'll fix this...

ig 2021-05-09