feat: Create 3 bubbles menacingly circling w/ color tint cosine palette
I called it 'A Kettle of Bubbles' because I thought the bubbles were
circling like vultures and a group of flying vultures is called a
'kettle'.

I've done this movement pattern (spinning shapes around a center) many
times but have never thought of them as vultures to my recollection. I
assume this is true now because of the noise distortion of the circles'
radii.

ig 2021-07-18