feat: Investigate dispersion algorithm & clean up its code while playing
It seemed that the hue color of the dispersion algorithm was the most
impactful on the end result. This might be surprising as because I just
realized that the dispersion "hue" color is being multiplied by the
scene color and not the other way around. The "hue" should tint the
scene just in the way the colors that come from a given refraction index
will be the scene color tinted by that color that corresponds to the
index.

ig 2021-08-29