feat: Experiment w/ additive component color palette in dispersion
Not sure if 'Additive Component' is the correct name for that style of
color palette. I can't remember what I called it before...

I also realized that my commit message last night was incorrect in
talking about the order being wrong when I multiplied the scene color
with the dispersion hue. Being multiplication it's commutative.

Decided to call it empty set because the new colors make it feel weirdly
empty.

ig 2021-08-30