feat: Play more w/ DOF & Fractals
My key finding was that I can push the dofCoeficient much further than I
have been. I've just been working on scales that lower values still made
an impact.

I also realized that the doFDistance is likely just whatever the
distance from the camera to the point in space that I want in focus is.
So really just shorter than the distance from the camera to the origin.

ig 2022-02-28