feat: Use goopy trail w/ a cube & dispersion
Did some thought work on making the trail adapt to the path given to
make it something you can just apply (assuming maybe being able to take
the derivative of the path etc), but I hit a wall both thought-wise and
energy-wise. I left the thoughts as a comment for future me or someone
smarter.

ig 2022-03-12