feat: Playing w/ periodic functions for domain warp
Figured out how to make the sigmoid / logistic function loop, but there
seems to be a discontinuity or extreme as the shape dips quite a bit.

I assume this trench is created from the way I force the input to loop
with a triangle wave. I think this happens because it also occurs when
using easing functions (another idea for periodic domain warp
functions).

ig 2022-06-04