feat: Experiment w/ changing warp direction by position
It didn't result in smooth be delineated regions of warping in one
direction vs another region, but it did give some characteristic of it.
One thing I'd like to try in the future is to have the warping based on
the angle in the xz plane to make the regions more discreet. This should
add motion internal to the region making it clearer when that direction
changes in the neighboring region.

ig 2022-08-08