feat: Recreate a manual holo effect
This holo effect created by marching a fix step size 'into' the shape
and accumulating noise as you go. In this case I tinted the noise with a
color based on the step count and accumulated by mixing between the
current accumulation and the current color based on the noise value.

ig 2022-08-12