feat: Create a colorful but mostly blue dispersion material
Returning to a color palette technique which mixes orthogonal or
component vectors based on individual indexes. The components can be
shifted or rotated in the same way to create a less common basis for the
color space. Today I also rotate the components by y position
effectively transitioning the color space across the object.

ig 2022-09-23