feat: Extrude a word svg & warp it
I believe I fixed the feedback bug that was somewhat happening
yesterday. At first I disabled the SVG texture loading today and thats
when it came back. Turns out I was forcing the SVG to be enabled but
didn't have it setup to use the default texture even with the SVG
texture enabled.

ig 2023-01-01