feat: Crop a word from itself and turn it into a mobius strip
Tried with a couple words and some work better than others. I especially
wanted 'Möbius' to work, but with the umlauts it was ugly with shading.
They really broke up the 'o' part when looking down on what would be on
the top side from the other cut.

ig 2023-01-13