feat: Create a school of gems
Experimented with abstracting gyroids to include other periodic
functions. I primarily used the triangle function but also realized that
I could use any function with a domain of [0, 1] if I mirror it so the
input goes 0 -> 1 -> 0. Then pass the input to the [0, 1] function and
use this as the 'metric' of the gyroid. (Metric is undoubtedly the wrong
term here as its not exactly determining the distance to the surface.)
Since gyroids use cosine / sine to determine their distance, I figured I
could just replace them with another function that has a range of [-1,
1].

ig 2023-02-02