feat: Mix dispersion color into background color w/ lambertian mask
I really like how the dispersion color combines with the base color
using a mix instead of addition or assignment. It allows it to be bold
and is a great perception mechanic when the base color is the same as
the background. Only the dispersion color gives the object form.

ig 2023-02-03