feat: Create a vibrant neon dispersion material
I initially tried to create a grid of overlapping squares which inverted
one another so that I could then shift each square with a delay based on
their starting position.

But, I couldn't figure it out. So Instead I did a quick experiment with
dispersion and was pleasantly surprised with what I stumbled upon!

ig 2023-03-30