feat: Attempt to use quadrant index to offset X in mobius space
It didn't work because there is a seam at -1 & 1 where the mobius
connects with itself. I tried lerping the amount of offset to 0 when it
gets close to the seam and rotating the quadrant index space to match
the other end (to smoothly transition). But neither worked as the had
down sides and artifacts still.

So instead I hid the seam with the camera! \(^.^ )/

ig 2023-06-09