feat: Nest "bin"s in each other delving infinitely deeper
Scaled the radius of the bins to create the effect that the camera is
zooming in with less issues with easing.

Fixed sdBin so that the opening is correctly sized with the back wall
being `thickness` thick.

Tried adding noise (and added support for mPos that is correct despite
all the scaling being done), but the noise wasn't as interesting as the
flat faces of the bin. Maybe useful for a different shape or scenario
that nests infinitely.

ig 2023-07-17