feat: Explore making only one object glow
I ended up using the new and improved bloom a bit, with a heavy dose of
the glow effect. The trick was trying to isolate the glow effect to just
the one object. I did this by using the trap parameter in the march
result and setting all objects' trap to be the distance to the object
that will glow. Now that the glow input was distance based and not based
on max iterations, I needed to adjust the scaling.

ig 2023-09-13