feat: Create a series of cosines that get out of phase
This session was about trying to learn new things especially around
curves. I started learning a ton about bezier curves and how to use
interpolation to construct them. I wanted to learn how to get the
distance to the curve however. IQ has an SDF for a quadric bezier, but I
couldn't understand it. Trying not to just copy and paste without
understanding, I moved to trying to find the SDF for a cosine (which was
my original goal actually). I started looking up SDFs for cosine and
found a few articles but they used approximations instead of analytic
methods for finding the distance. Finding IQ also had an approximate
method, I decided to just learn this algorithm instead of trying to
figure out an analytic solution.

The cosine distance can be normalized to y(x) = w * cos(x) as IQ notes
by offsetting by the y offset and the phase (aka the x offset). The
frequency and amplitude can be combined into a single term `w` by
scaling the input vector by the frequency and then scale the amplitude
by the frequency to keep it to scale. With the reduced equation, it
becomes a minimizing step where we try to find an `x` that sets the
derivative of `y` to `0`. IQ then uses a bisection search with enough
iterations to ensure floating point accuracy making the answer
equivalent to an analytic method.

Finally I also employed the `fwidth(n)` trick from The Art of Code for
antialiasing my edges. I applied this across the fragment shader hoping
that it applied everywhere that I was using `edge`. I was lazy and did
leave some `edge`s in the commented out sections.

ig 2024-02-07