feat: Create heavily offset layers that shift the wave
Was stuck on issues with the loop jumping, but eventually figured out
that my shift to the vec2 domain for the cosine SDF was not scaled on
the x axis correctly. I wasn't thinking about how the input is scaled
via the frequency, so the shift only needs to be [0, 1).

ig 2024-02-08